requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3

# git init 
# git remote add origin ссылка
# git add
# git commit -m 'commit'
# git push origin master 



# git add
# git commit -m 'commit'
# git push origin master 
